# marketing-contracts/src/main/resources/openapi/marketing-api.yaml
openapi: 3.1.0
info:
  title: Marketing Low-code Platform API
  version: 1.0.0
  description: R1 edge contract. Tenant and organization scope come from the verified identity in OIDC mode.
servers:
  - {url: http://localhost:8080, description: Local edge gateway}
security: [{oidc: []}]
...
paths:
...
  /api/v1/fields:
    get: {tags: [Audience], operationId: listAudienceFields, responses: *list}
    post: {tags: [Audience], operationId: registerAudienceField, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *created}
  /api/v1/audiences:
    get: {tags: [Audience], operationId: listAudiences, responses: *list}
    post: {tags: [Audience], operationId: createAudience, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *created}
  /api/v1/audiences/{segmentId}/versions/{version}:preview:
    post:
      tags: [Audience]
      operationId: previewAudience
      parameters: [{$ref: '#/components/parameters/SegmentId'}, {$ref: '#/components/parameters/Version'}]
      requestBody: {$ref: '#/components/requestBodies/JsonArray'}
      responses: *ok
  /api/v1/audiences/{segmentId}/versions/{version}/snapshots:
    post:
      tags: [Audience]
      operationId: createAudienceSnapshot
      parameters: [{$ref: '#/components/parameters/SegmentId'}, {$ref: '#/components/parameters/Version'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *created
  /api/v1/audiences/snapshots/{snapshotId}/memberships:
    put: {tags: [Audience], operationId: updateAudienceMembership, parameters: [{$ref: '#/components/parameters/SnapshotId'}], requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/audiences/snapshots/{snapshotId}/memberships/{subjectToken}:
    get:
      tags: [Audience]
      operationId: getAudienceMembership
      parameters:
        - {$ref: '#/components/parameters/SnapshotId'}
        - {$ref: '#/components/parameters/SubjectToken'}
        - {in: query, name: usedAt, required: true, schema: {type: string, format: date-time}}
        - {in: query, name: stalePolicy, required: true, schema: {type: string, enum: [REJECT_CANDIDATE, USE_LAST_GOOD, GENERIC_PATH]}}
      responses: *ok
...
  /api/v1/funding/accounts:
    get: {tags: [Benefit], operationId: listFundingAccounts, responses: *list}
    post: {tags: [Benefit], operationId: createFundingAccount, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *created}
  /api/v1/benefits:
    get: {tags: [Benefit], operationId: listBenefits, responses: *list}
  /api/v1/benefits/{benefitId}:
    get:
      tags: [Benefit]
      operationId: getBenefit
      parameters: [{$ref: '#/components/parameters/BenefitId'}]
      responses: *ok
    put:
      tags: [Benefit]
      operationId: putBenefit
      parameters: [{$ref: '#/components/parameters/BenefitId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
  /api/v1/funding/accounts/{resourceKey}:advance-fence:
    post: {tags: [Benefit], operationId: advanceFundingFence, parameters: [{$ref: '#/components/parameters/ResourceKey'}], requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/promotion-applications:
    post:
      tags: [Benefit]
      operationId: reservePromotion
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/ReservePromotion'}
      responses: *created
  /api/v1/promotion-applications/{applicationId}:confirm:
    post: {tags: [Benefit], operationId: confirmPromotion, parameters: [{$ref: '#/components/parameters/ApplicationId'}, {$ref: '#/components/parameters/IdempotencyKey'}], requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/promotion-applications/{applicationId}:cancel:
    post: {tags: [Benefit], operationId: cancelPromotion, parameters: [{$ref: '#/components/parameters/ApplicationId'}, {$ref: '#/components/parameters/IdempotencyKey'}], requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/promotion-applications/{applicationId}:refund:
    post: {tags: [Benefit], operationId: refundPromotion, parameters: [{$ref: '#/components/parameters/ApplicationId'}, {$ref: '#/components/parameters/IdempotencyKey'}], requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/promotion-applications/{applicationId}:reverse:
    post: {tags: [Benefit], operationId: reversePromotion, parameters: [{$ref: '#/components/parameters/ApplicationId'}, {$ref: '#/components/parameters/IdempotencyKey'}], requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/promotion-applications:expire:
    post: {tags: [Benefit], operationId: expirePromotions, parameters: [{in: query, name: limit, required: false, schema: {type: integer, default: 100, minimum: 1}}], responses: *ok}
  /api/v1/funding/reconciliation:
    get: {tags: [Benefit], operationId: reconcileFunding, responses: *ok}
  /api/v1/events/sources:
    post: {tags: [Event], operationId: registerEventSource, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *created}
  /api/v1/events:
    post: {tags: [Event], operationId: ingestEvent, requestBody: {$ref: '#/components/requestBodies/InboundEvent'}, responses: *accepted}
  /api/v1/quarantine:
    get: {tags: [Event], operationId: listQuarantine, responses: *list}
  /api/v1/quarantine/{quarantineId}:replay:
    post: {tags: [Event], operationId: replayQuarantinedEvent, parameters: [{$ref: '#/components/parameters/QuarantineId'}], responses: *ok}
...
  /api/v1/consents:
    put: {tags: [Engagement], operationId: setConsent, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/consents/suppressions:
    put: {tags: [Engagement], operationId: setSuppression, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/contacts/frequency-policies:
    put: {tags: [Engagement], operationId: setFrequencyPolicy, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/templates:
    get: {tags: [Engagement], operationId: listTemplates, responses: *list}
    post: {tags: [Engagement], operationId: createTemplate, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *created}
  /api/v1/contacts:
    get:
      tags: [Engagement]
      operationId: listContacts
      parameters:
        - {in: query, name: state, required: false, schema: {type: string, enum: [PENDING, RETRY_PENDING, ACCEPTED, SENT, DELIVERED, FAILED, CLICKED, UNSUBSCRIBED]}}
        - {in: query, name: limit, required: false, schema: {type: integer, default: 50, minimum: 1, maximum: 200}}
      responses: *list
    post: {tags: [Engagement], operationId: sendContact, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *accepted}
  /api/v1/contacts/{contactKey}:
    get: {tags: [Engagement], operationId: getContact, parameters: [{$ref: '#/components/parameters/ContactKey'}], responses: *ok}
  /api/v1/providers/callbacks:
    post:
      tags: [Engagement]
      operationId: receiveProviderCallback
      parameters:
        - {in: header, name: X-Marketing-Provider-Request-Id, required: true, schema: {type: string}}
        - {in: header, name: X-Marketing-Timestamp, required: true, schema: {type: string, format: date-time}}
        - {in: header, name: X-Marketing-Signature, required: true, schema: {type: string, pattern: '^[a-f0-9]{64}$'}}
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
...
components:
  securitySchemes:
    oidc: {type: openIdConnect, openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration}
  parameters:
    IdempotencyKey: {in: header, name: Idempotency-Key, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
...
    Version: {in: path, name: version, required: true, schema: {type: integer, format: int64, minimum: 1}}
...
    SegmentId: {in: path, name: segmentId, required: true, schema: {type: string}}
    SnapshotId: {in: path, name: snapshotId, required: true, schema: {type: string}}
    SubjectToken: {in: path, name: subjectToken, required: true, schema: {type: string}}
    ResourceKey: {in: path, name: resourceKey, required: true, schema: {type: string}}
    BenefitId: {in: path, name: benefitId, required: true, schema: {type: string}}
    ApplicationId: {in: path, name: applicationId, required: true, schema: {type: string}}
    QuarantineId: {in: path, name: quarantineId, required: true, schema: {type: string}}
...
    ContactKey: {in: path, name: contactKey, required: true, schema: {type: string, minLength: 1}}
...
  requestBodies:
    JsonObject: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/JsonObject'}}}}
    JsonArray: {required: true, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/JsonObject'}}}}}
...
    ReservePromotion: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReservePromotionRequest'}}}}
    InboundEvent: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/InboundEvent'}}}}
...
  schemas:
    JsonObject: {type: object, additionalProperties: true}
...
    ReservePromotionRequest:
      type: object
      additionalProperties: false
      required: [offerToken, cartDigest, expectedFencingEpochs]
      properties: {offerToken: {type: string}, cartDigest: {type: string}, orderId: {type: string}, expectedFencingEpochs: {type: object, additionalProperties: {type: integer, format: int64, minimum: 1}}}
    InboundEvent:
      type: object
      additionalProperties: false
      required: [eventId, sourceId, eventType, businessKey, subjectToken, occurredAt, schemaVersion, data]
      properties:
        eventId: {type: string}
        sourceId: {type: string}
        eventType: {type: string}
        businessKey: {type: string}
        subjectToken: {type: string}
        occurredAt: {type: string, format: date-time}
        schemaVersion: {type: string}
        aggregateVersion: {type: [integer, 'null'], format: int64}
        data: {type: object, additionalProperties: true}
...